+++ apps/storybook/stories/button/ButtonIconLink.stories.tsx
@@ -163,21 +163,6 @@ export default {
         required: false,
       },
     },
-    iconColorHex: {
-      description: 'Text color of the button (hex)',
-      type: {
-        name: 'string',
-        required: false,
-      },
-    },
-    iconColor: {
-      description: 'Text color of the button',
-      type: {
-        name: 'enum',
-        value: Object.keys(Colors),
-        required: false,
-      },
-    },
   },
 } as Meta<ButtonIconLinkProps>