+++ packages/button/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @pikas-ui/button
 
+## 1.1.0
+
+### Minor Changes
+
+- Update color system
+
+### Patch Changes
+
+- Updated dependencies [abe30ab]
+  - @pikas-ui/styles@1.1.0
+  - @pikas-ui/icons@1.0.26
+  - @pikas-ui/loader@1.0.26
+
 ## 1.0.28
 
 ### Patch Changes