+++ packages/slider/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pikas-ui/slider
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [abe30ab]
+  - @pikas-ui/styles@1.1.0
+  - @pikas-ui/icons@1.0.26
+  - @pikas-ui/text@1.0.27
+
 ## 1.0.2
 
 ### Patch Changes